+++ images/media/imgopt.sh
@@ -16,4 +16,5 @@
 # limitations under the License.
 
 
-find -type f -name "*.jpg" -exec jpegoptim -f --overwrite --max=90 --strip-all {} \;
+find -type f -name "*.jpg" -exec jpegoptim -f --max=80 --strip-all {} \;
+for f in *.jpg; do convert "$f" -resize '512x>' "$f"; done  # resize image to 512 pixels wide