+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 ## 0.5.7 - 2017-03-01
 
 ### Added
-- quotes carousel on landing page
+- quotes carousel on landing page, why support us page
 
 ## 0.5.6 - 2017-02-28