+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 0.5.8 - 2017-03-02
+
+### Fixed
+- too small chart labels
+
 ## 0.5.7 - 2017-03-01
 
 ### Added